+++ apps/backoffice/CHANGELOG.md
@@ -1,5 +1,18 @@
 # io-services-cms-backoffice
 
+## 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [67d6b8ef]
+  - @io-services-cms/external-clients@1.11.0
+
+## 2.4.0
+
+### Minor Changes
+
+- 9ade29a5: add api keys exports file generation use case
+
 ## 2.3.0
 
 ### Minor Changes